requirements.txt
aiomysql==0.1.1
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.3.0
bcrypt==4.1.1
boto3==1.34.54
botocore==1.34.137
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
dynaconf==3.2.4
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
mailtrap==2.0.1
Mako==1.3.0
MarkupSafe==2.1.3
mypy==1.6.1
mypy-extensions==1.0.0
orjson==3.9.10
passlib==1.7.4
pip==24.1.1
psycopg==3.1.14
psycopg-binary==3.1.14
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.2
setuptools==58.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
types-requests==2.32.0.20240712
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.2.2
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
websockets==12.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
async-timeout==4.0.3
attrs==24.2.0
multidict==6.0.5